+++ src/scene/transform.h
@@ -1,9 +1,9 @@
 #pragma once
 
-inline glm::dvec3 operator*(const glm::dmat4x4 &mat, const glm::dvec3 &vec) {
-    glm::dvec4 vec4(vec[0], vec[1], vec[2], 1.0);
+inline glm::vec3 operator*(const glm::mat4x4 &mat, const glm::vec3 &vec) {
+    glm::vec4 vec4(vec[0], vec[1], vec[2], 1.0);
     auto ret = mat * vec4;
-    return glm::dvec3(ret[0], ret[1], ret[2]);
+    return glm::vec3(ret[0], ret[1], ret[2]);
 }
 
 class TransformNode {
@@ -13,40 +13,40 @@ public:
             delete c;
     }
 
-    TransformNode *createChild(const glm::dmat4x4 &txf) {
+    TransformNode *createChild(const glm::mat4x4 &txf) {
         auto child = new TransformNode(this, txf);
         children.push_back(child);
         return child;
     }
 
-    glm::dvec3 localToGlobalCoords(const glm::dvec3 &v) {
+    glm::vec3 localToGlobalCoords(const glm::vec3 &v) {
         return xform * v;
     }
 
-    glm::dvec4 localToGlobalCoords(const glm::dvec4 &v) {
+    glm::vec4 localToGlobalCoords(const glm::vec4 &v) {
         return xform * v;
     }
 
-    glm::dvec3 localToGlobalCoordsNormal(const glm::dvec3 &v) {
+    glm::vec3 localToGlobalCoordsNormal(const glm::vec3 &v) {
         return glm::normalize(normi * v);
     }
 
     const glm::dmat4x4 &transform() const { return xform; }
 
 protected:
-    TransformNode(TransformNode *parent, const glm::dmat4x4 &xform) {
+    TransformNode(TransformNode *parent, const glm::mat4x4 &xform) {
         this->parent = parent;
         if (parent == nullptr)
             this->xform = xform;
         else
             this->xform = parent->xform * xform;
         inverse = glm::inverse(this->xform);
-        normi = glm::transpose(glm::inverse(glm::dmat3x3(this->xform)));
+        normi = glm::transpose(glm::inverse(glm::mat3x3(this->xform)));
     }
 
-    glm::dmat4x4 xform;
-    glm::dmat4x4 inverse;
-    glm::dmat3x3 normi;
+    glm::mat4x4 xform;
+    glm::mat4x4 inverse;
+    glm::mat3x3 normi;
 
     TransformNode *parent;
     std::vector<TransformNode *> children;
@@ -54,5 +54,5 @@ protected:
 
 class TransformRoot : public TransformNode {
 public:
-    TransformRoot() : TransformNode(nullptr, glm::dmat4x4(1.0)) {}
+    TransformRoot() : TransformNode(nullptr, glm::mat4x4(1.0)) {}
 };
